+++ application/controller/404.php
@@ -1,1 +1,4 @@
-<?php header("HTTP/1.0 404 Not Found"); ?>
+<?php
+	$app->setTitle("Not found");
+	header("HTTP/1.0 404 Not Found");
+?>